Modificato microservizio craftsman

+++ craftsman-service/src/main/java/com/unito/edu/craftsmanservice/repository/CraftsmanRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CraftsmanRepository extends JpaRepository<Craftsman,Long> {
+public interface CraftsmanRepository extends JpaRepository<Craftsman,Integer> {
 }
